Refactor: 컴포넌트 분리 및 리펙토링 작업
