+++ app/(route)/activity/details/[id]/page.tsx
