+++ app/(route)/activity/details/_components/Calendar/MobilePeopleCounter.tsx
