+++ app/(route)/activity/details/_components/Calendar/PeopleCounter.tsx
