+++ app/(route)/activity/details/_components/ExperienceDetail/types.ts
@@ -1,0 +1,32 @@
+export interface Review {
+  content: string;
+  id: number;
+  rating: number;
+  user: {
+    id: number;
+    nickname: string;
+  };
+}
+
+export interface ExperienceDetailProps {
+  averageRating: number;
+  experience: {
+    address: string;
+    bannerImageUrl: string;
+    category: string;
+    createdAt: string;
+    description: string;
+    id: number;
+    price: number;
+    rating: number;
+    reviewCount: number;
+    schedules: { date: string; endTime: string; id: number; startTime: string }[];
+    subImages: { id: number; imageUrl: string }[];
+    title: string;
+    updatedAt: string;
+    userId: number;
+  };
+  reviews: Review[];
+  totalReviews: number;
+  currentUserId: number | null;
+}
