+++ app/(route)/activity/details/_components/NaverMap/index.tsx
