+++ app/_apis/activities/fetchReviewsData.ts
@@ -1,0 +1,11 @@
+import axiosInstance from '@/_libs/axios';
+import { ReviewResponse } from '@/_types/details/types';
+
+export const fetchReviewsData = async (activityId: string): Promise<ReviewResponse> => {
+  try {
+    const response = await axiosInstance.get<ReviewResponse>(`/activities/${activityId}/reviews`);
+    return response.data;
+  } catch (err) {
+    throw new Error('리뷰 데이터를 불러오는 중 오류가 발생했습니다.');
+  }
+};
